src/requirements.txt
python-telegram-bot[job-queue]
python-telegram-bot
nest-asyncio
requests
beautifulsoup4
